requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.5.0; python_version >= '3.9'
aiohttp==3.11.13; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
attrs==25.1.0; python_version >= '3.8'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
frozenlist==1.5.0; python_version >= '3.8'
greenlet==3.1.1; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
langchain==0.0.352; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-community==0.0.20; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-core==0.1.23; python_full_version >= '3.8.1' and python_version < '4.0'
langsmith==0.0.87; python_full_version >= '3.8.1' and python_version < '4.0'
marshmallow==3.26.1; python_version >= '3.9'
multidict==6.1.0; python_version >= '3.8'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.4; python_version >= '3.9'
openai==0.27.8; python_full_version >= '3.7.1'
packaging==23.2; python_version >= '3.7'
propcache==0.3.0; python_version >= '3.9'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
python-dotenv==1.0.0; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.38; python_version >= '3.7'
tenacity==8.5.0; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
urllib3==2.3.0; python_version >= '3.9'
yarl==1.18.3; python_version >= '3.9'
